requirements.txt
aiohttp==3.7.4.post0
aiosignal==1.3.1
async-timeout==3.0.1
attrs==22.1.0
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.1
discord-ext-audiorec==0.1.6
discord.py==1.7.3
frozenlist==1.3.3
idna==3.4
load-dotenv==0.1.0
multidict==6.0.3
pycparser==2.21
PyNaCl==1.4.0
python-dotenv==0.21.0
six==1.16.0
typing_extensions==4.4.0
yarl==1.8.2
